refactor: move user chat prompt to experimental routes
Move the user-level custom prompt endpoints from /api/v2/users/{user}/
to /api/experimental/chats/config/ alongside other chat configuration:

- GET /api/experimental/chats/config/user-prompt
- PUT /api/experimental/chats/config/user-prompt

Changes:
- Handlers moved from coderd/users.go to coderd/chats.go
- Use httpmw.APIKey(r).UserID instead of httpmw.UserParam(r)
- SDK types + client methods moved from codersdk/users.go to
  codersdk/chats.go
- Routes registered under /chats/config alongside system-prompt
- Swagger annotations removed (experimental endpoints are hidden
  from docs gen)
